update and add test
